Separate functionality of scripts from invocation. Capacity check can trigger builds. Triggering a build can start watching the build.
